feat(ssg): decoupling generate_html
