+++ src/utilities.rs
@@ -1,11 +1,13 @@
 // Copyright © 2023 Shokunin (職人) Static Site Generator. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+extern crate regex;
 use minify_html::{minify, Cfg};
 use quick_xml::{
     events::{BytesEnd, BytesStart, BytesText, Event},
     Writer,
 };
+use regex::Regex;
 use std::{
     borrow::Cow,
     error::Error,
@@ -356,4 +358,127 @@ pub fn to_title_case(s: &str) -> String {
         })
         .collect::<Vec<_>>()
         .join(" ")
+}
+
+/// Formats a header string with an ID and class attribute.
+pub fn format_header_with_id_class(
+    header_str: &str,
+    id_regex: &Regex,
+) -> String {
+    let mut formatted_header_str = String::new();
+    let mut in_header_tag = false;
+    let mut id_attribute_added = false;
+    let mut class_attribute_added = false;
+
+    for c in header_str.chars() {
+        if !in_header_tag {
+            formatted_header_str.push(c);
+            if c == '<' {
+                in_header_tag = true;
+            }
+        } else {
+            if !id_attribute_added && c == ' ' {
+                formatted_header_str.push_str(&format!(
+                    " id=\"{}\"",
+                    id_regex
+                        .replace_all(&header_str.to_lowercase(), "-")
+                        .trim_matches('-')
+                        .trim_end_matches("-h1")
+                        .trim_end_matches("-h2")
+                        .trim_end_matches("-h3")
+                        .trim_end_matches("-h4")
+                        .trim_end_matches("-h5")
+                        .trim_end_matches("-h6")
+                ));
+                id_attribute_added = true;
+            }
+            if !class_attribute_added && c == '>' {
+                formatted_header_str.push_str(&format!(
+                    " class=\"{}\"",
+                    id_regex
+                        .replace_all(&header_str.to_lowercase(), "-")
+                        .trim_matches('-')
+                        .trim_end_matches("-h1")
+                        .trim_end_matches("-h2")
+                        .trim_end_matches("-h3")
+                        .trim_end_matches("-h4")
+                        .trim_end_matches("-h5")
+                        .trim_end_matches("-h6")
+                ));
+                class_attribute_added = true;
+            }
+            formatted_header_str.push(c);
+            if c == '>' {
+                in_header_tag = false;
+            }
+        }
+    }
+    formatted_header_str
+}
+
+/// Extracts the front matter from the given content.
+pub fn extract_front_matter(content: &str) -> &str {
+    if content.starts_with("---\n") {
+        if let Some(end_pos) = content.find("\n---\n") {
+            &content[end_pos + 5..] // Skip the "---\n\n" that follows the front matter
+        } else {
+            ""
+        }
+    } else if content.starts_with("+++\n") {
+        if let Some(end_pos) = content.find("\n+++\n") {
+            &content[end_pos + 5..] // Skip the "+++\n\n" that follows the front matter
+        } else {
+            ""
+        }
+    } else if content.starts_with("{\n") {
+        if let Some(end_pos) = content.find("\n}\n") {
+            &content[end_pos + 2..]
+        } else {
+            ""
+        }
+    } else {
+        content
+    }
+}
+
+/// Creates the default Comrak options.
+pub fn create_comrak_options() -> comrak::ComrakOptions {
+    let mut options = comrak::ComrakOptions::default();
+
+    // Enable non-standard Markdown features:
+    options.extension.autolink = true; // Detects URLs and email addresses and makes them clickable.
+    options.extension.description_lists = true; // Allows you to create description lists.
+    options.extension.footnotes = true; // Allows you to create footnotes.
+    options.extension.front_matter_delimiter = Some("---".to_owned()); // Ignore front-mater starting with '---'
+    options.extension.header_ids = Some("".to_string()); // Adds an ID to each header.
+    options.extension.strikethrough = true; // Allows you to create strikethrough text.
+    options.extension.superscript = true; // Allows you to create superscript text.
+    options.extension.table = true; // Allows you to create tables.
+    options.extension.tagfilter = true; // Allows you to filter HTML tags.
+    options.extension.tasklist = true; // Allows you to create task lists.
+    options.parse.smart = true; // Enables smart punctuation.
+    options.render.github_pre_lang = true; // Renders GitHub-style fenced code blocks.
+    options.render.hardbreaks = false; // Renders hard line breaks as <br> tags.
+    options.render.unsafe_ = true; // Allows raw HTML to be rendered.
+
+    options
+}
+
+/// Updates the class attributes in a line of content
+pub fn update_class_attributes(
+    line: &str,
+    class_regex: &Regex,
+    img_regex: &Regex,
+) -> String {
+    if line.contains(".class=&quot;") {
+        let captures = class_regex.captures(line).unwrap();
+        let class_value = captures.get(1).unwrap().as_str();
+        let updated_line = class_regex.replace(line, "");
+        let updated_line_with_class = img_regex.replace(
+            &updated_line,
+            &format!("$1 class=\"{}\"$2", class_value),
+        );
+        return updated_line_with_class.into_owned();
+    }
+    line.to_owned()
 }
